feat: 메인 페이지에서 참여코드 입력 페이지 구현
